requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.33
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
Werkzeug==2.3.7
pillow==10.0.1
python-dotenv==1.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2